+++ CHANGELOG.md
@@ -11,10 +11,10 @@ This project adheres to [Semantic Versioning](http://semver.org/).
   - text from first `h1` is injected into `title`
 
 ## [1.1.0]
-- avoid rewriting external links to MDs ([#3](https://github.com/ngoldman/sitedown/pull/3))
-- generate CSS classes for highlightjs support ([#4](https://github.com/ngoldman/sitedown/pull/4))
+- avoid rewriting external links to MDs ([#3](https://github.com/ungoldman/sitedown/pull/3))
+- generate CSS classes for highlightjs support ([#4](https://github.com/ungoldman/sitedown/pull/4))
 
 ## 1.0.0
 - engage
 
-[1.1.0]: https://github.com/ngoldman/sitedown/compare/v1.0.0...v1.1.1
+[1.1.0]: https://github.com/ungoldman/sitedown/compare/v1.0.0...v1.1.1